+++ CHANGES.md
@@ -7,7 +7,7 @@ Version 0.1.2
 To be released.
 
  -  Updated the *Standard Korean Language Dictionary* data
-    (*data/ko-kr-stdict.tsv*) to the revision October 2019.
+    (*data/ko-kr-stdict.tsv*) to the revision 2019-10-14.
 
 
 Version 0.1.1