+++ CHANGES.md
@@ -1,7 +1,7 @@
 Seonbi changelog
 ================
 
-Version 0.1.2
+Version 0.2.0
 -------------
 
 To be released.